Move OffsetBuilder into des_meds.py as a helper builder. (#928)
