+++ tests/test_des.py
@@ -256,7 +256,7 @@ def test_meds_config():
     # Some parameters:
     if __name__ == '__main__':
         nobj = 5
-        n_per_obj = 3
+        n_per_obj = 8
     else:
         nobj = 5
         n_per_obj = 3
@@ -267,9 +267,8 @@ def test_meds_config():
     g1 = -0.17
     g2 = 0.23
 
-    #generate offsets that depend on the 
-    #on the object num so can be easily reproduced
-    #for testing below
+    # generate offsets that depend on the on the object num so can be easily reproduced
+    # for testing below
     offset_x = '$ np.sin(999.*(@obj_num+1))'
     offset_y = '$ np.sin(998.*(@obj_num+1))'
     def get_offset(obj_num):
@@ -380,23 +379,23 @@ def test_meds_config():
 
             # cutout_row and cutout_col are the "zero-offset"
             # position of the object in the stamp. In this convention, the center
-            # of the first pixel is at (0,0), call this meds_center. 
-            #This means cutout_row/col should be the same as meds_center + offset
+            # of the first pixel is at (0,0), call this meds_center.
+            # This means cutout_row/col should be the same as meds_center + offset
             offset = get_offset(iobj*n_cut+icut)
             meds_center = galsim.PositionD( (box_size-1.)/2., (box_size-1.)/2. )
             cutout_row = cat['cutout_row'][iobj][icut]
             cutout_col = cat['cutout_col'][iobj][icut]
             print('cutout_row, cutout_col = ',cutout_col, cutout_row)
-            numpy.testing.assert_almost_equal(cutout_col, 
+            numpy.testing.assert_almost_equal(cutout_col,
                                               (meds_center+offset).x)
-            numpy.testing.assert_almost_equal(cutout_row, 
+            numpy.testing.assert_almost_equal(cutout_row,
                                               (meds_center+offset).y)
 
             # The col0 and row0 here should be the same.
             wcs_meds = m.get_jacobian(iobj, icut)
-            numpy.testing.assert_almost_equal(wcs_meds['col0'], 
+            numpy.testing.assert_almost_equal(wcs_meds['col0'],
                                               (meds_center+offset).x)
-            numpy.testing.assert_almost_equal(wcs_meds['row0'], 
+            numpy.testing.assert_almost_equal(wcs_meds['row0'],
                                               (meds_center+offset).y)
 
 
@@ -409,9 +408,9 @@ def test_meds_config():
             print('centroid = ',ix/itot, iy/itot)
 
             print('center + offset = ',meds_center + offset)
-            numpy.testing.assert_almost_equal(ix/itot, 
+            numpy.testing.assert_almost_equal(ix/itot,
                                               (meds_center+offset).x, decimal=2)
-            numpy.testing.assert_almost_equal(iy/itot, 
+            numpy.testing.assert_almost_equal(iy/itot,
                                               (meds_center+offset).y, decimal=2)
 
             # The orig positions are irrelevant and should be 0.
